Fix samples for 64-bit ABIs
1. Set APP_ABI:=all in almost all samples
2. Enhance hello-jni to report more ABIs
3. Fix warnings

+++ native-plasma/jni/plasma.c
@@ -224,7 +224,7 @@ static void fill_plasma(ANativeWindow_Buffer* buffer, double  t)
         uint16_t*  line_end = line + buffer->width;
 
         if (line < line_end) {
-            if (((uint32_t)line & 3) != 0) {
+            if (((uint32_t)(uintptr_t)line & 3) != 0) {
                 Fixed ii = base + fixed_sin(xt1) + fixed_sin(xt2);
 
                 xt1 += XT1_INCR;
